refactor(fontmapper): extract shared classes
