+++ fontmapper/FM16/FMS6.py
@@ -1,15 +1,26 @@
+#!/usr/bin/env python3
+"""FontMapper FM16 utility functions."""
+from __future__ import annotations
+
+try:
+    from AGENTS.tools.header_utils import ENV_SETUP_BOX
+except Exception:
+    import sys
+    print(ENV_SETUP_BOX)
+    sys.exit(1)
+# --- END HEADER ---
+
 import json
 import numpy as np
 import os
 import random
 import torch
-import torch.nn as nn
 import torch.nn.functional as F
 import yaml
 from collections import defaultdict
 from fontTools.ttLib import TTFont
 from PIL import Image, ImageDraw, ImageFont, ImageOps, ImageChops
-from torch.utils.data import DataLoader, Dataset
+from torch.utils.data import DataLoader
 from torchvision import transforms
 from torchvision.transforms import functional as TF
 import io
@@ -19,6 +30,18 @@ from flask_cors import CORS
 import sys
 import platform
 import math
+
+from .modules import (
+    ModelCompatConfig,
+    ModelConfig,
+    CharSorter,
+    AddRandomNoise,
+    DistortionChain,
+    RandomGaussianBlur,
+    ToTensorAndToDevice,
+    CustomDataset,
+    CustomInputDataset,
+)
 
 # --- Add these flags at the top ---
 RUN_SERVER = False
@@ -102,89 +125,6 @@ print(globals())
 FMversion = "FM37"
 
 
-class ModelCompatConfig:
-    def __init__(self, charset, font_files, font_size, conv1_out, conv2_out, linear_out, width, height):
-        self.charset = charset
-        self.width = width
-        self.height = height
-        self.font_files = font_files
-        self.font_size = font_size
-        self.conv1_out = conv1_out
-        self.conv2_out = conv2_out
-        self.linear_out = linear_out
-        
-class ModelConfig:
-    def __init__(self, compatibility_model, dropout, learning_rate, epochs, batch_size, demo_image=None, epochs_per_preview=10000000, model_path="model"):
-        self.charset = compatibility_model.charset
-        self.refresh_rate = refresh_rate
-        self.demo_image = demo_image
-        self.demo_images = demo_images
-        self.training_categories = training_categories
-        self.training_gradients = training_gradients
-        self.training_images = training_images
-        self.human_in_the_loop = hitl
-        self.image_batch_limit = image_batch_limit
-        self.epochs_per_preview = epochs_per_preview
-        self.font_files = compatibility_model.font_files
-        self.font_size = compatibility_model.font_size
-        self.conv1_out = compatibility_model.conv1_out
-        self.conv2_out = compatibility_model.conv2_out
-        self.linear_out = compatibility_model.linear_out
-        self.width = compatibility_model.width
-        self.height = compatibility_model.height
-        self.dropout = dropout
-        self.learning_rate = learning_rate
-        self.epochs = epochs
-        self.batch_size = batch_size
-        self.model = None
-        self.dataloader = None
-        self.gradient_loss_function = gradient_loss_function
-        self.training_loss_function = training_loss_function
-        self.demo_loss_function = demo_loss_function
-        self.version = f"{font_size}-{conv1_out}-{conv2_out}-{linear_out}-{dropout}-{learning_rate}-prev-{model_base}"
-        self.model_path = f"{FMversion}--{self.version}--{model_name}"
-        if model_path_override != "":
-            self.model_path = model_path_override
-
-
-import inspect
-some_model_directory = "./"
-
-
-
-class CharSorter(nn.Module):
-    def __init__(self, config):
-        super(CharSorter, self).__init__()
-        self.config = config
-        num_classes = len(config.charset)
-        self.conv1 = nn.Conv2d(1, config.conv1_out, kernel_size=5, stride=1, padding=2)
-        self.pool = nn.MaxPool2d(kernel_size=2, stride=2, padding=0)
-        self.conv2 = nn.Conv2d(config.conv1_out, config.conv2_out, kernel_size=5, stride=1, padding=2)
-
-        # Dummy forward pass to determine size automatically
-        dummy_input = torch.autograd.Variable(torch.ones(1, 1, config.height, config.width))
-        output = self.conv2(self.pool(self.conv1(dummy_input)))
-        output = self.pool(output)
-        n_size = output.numel() // output.shape[0]
-
-        self.fc1 = nn.Linear(n_size, config.linear_out)
-        self.fc2 = nn.Linear(config.linear_out, num_classes)
-        self.relu = nn.ReLU()
-        self.dropout = nn.Dropout(p=config.dropout)
-        self.charset = config.charset
-        self.charBitmasks = config.charBitmasks
-        self.demo_width = config.width
-        self.demo_height = config.height
-
-    def forward(self, x):
-        x = x.view(-1, 1, self.config.height, self.config.width)
-        x = self.pool(self.relu(self.conv1(x)))
-        x = self.pool(self.relu(self.conv2(x)))
-        x = x.view(-1, self.fc1.in_features)
-        x = self.relu(self.fc1(x))
-        x = self.dropout(x)
-        x = self.fc2(x)
-        return x
     
 
 def list_printable_characters(font_path, font_size=12, epsilon=1):
@@ -395,57 +335,6 @@ def create_config(base_config, font_files=None, font_size=None, complexity_level
     )
 
 
-          
-class AddRandomNoise(object):
-    def __call__(self, img):
-        noise = torch.randn_like(img, device=device) * 0.1
-        return torch.clamp(img + noise, 0, 1)
-
-class DistortionChain(object):
-    def __init__(self, noise_chance=0.5, blur_chance=0.5):
-        self.noise_chance = noise_chance
-        self.blur_chance = blur_chance
-        self.noise_transform = AddRandomNoise()
-        self.blur_transform = RandomGaussianBlur()
-
-    def __call__(self, img):
-        if random.random() < self.noise_chance:
-            img = self.noise_transform(img)
-        if random.random() < self.blur_chance:
-            img = self.blur_transform(img)
-        return img
-class RandomGaussianBlur(object):
-    def __call__(self, img):
-        sigma = random.uniform(0.5, 1.5)
-        max_kernel_size = min(img.shape[-2], img.shape[-1]) // 10  # Limit kernel size to 10% of image size
-        kernel_size = 1 + 2 * random.randint(0, max_kernel_size+1)  # Randomly select kernel size within the limit
-        img = TF.gaussian_blur(img, kernel_size=(kernel_size, kernel_size), sigma=sigma)
-        return img
-class ToTensorAndToDevice:
-    def __init__(self, device):
-        self.device = device
-
-    def __call__(self, pic):
-        tensor = transforms.ToTensor()(pic)
-        tensor = tensor.to(torch.float16)
-        tensor = tensor.to(self.device)
-        return tensor
-    
-class CustomDataset(Dataset):
-    def __init__(self, char_bitmasks, labels, transform=None):
-        self.char_bitmasks = [Image.fromarray(obj).convert('L') for obj in char_bitmasks]
-        self.labels = labels.clone()
-        self.transform = transform
-
-    def __getitem__(self, idx):
-        image = self.char_bitmasks[idx]
-        label = self.labels[idx]
-        if self.transform:
-            image = self.transform(image)
-        return image, label.clone()
-
-    def __len__(self):
-        return len(self.char_bitmasks)
     
 
 
@@ -608,34 +497,6 @@ def slice_and_separate_channels(image, chunk_size, image_id):
             slice_id += 1
     return separated_channels
 
-class CustomInputDataset(Dataset):
-    def __init__(self, tensor_images, network_size, mode='auto', transform=None):
-        self.network_size = network_size
-        self.tensor_images = tensor_images
-        self.transform = transform
-        self.prepared_data = self.prepare_data(tensor_images, network_size, mode)
-        
-
-    def prepare_data(self, tensor_images, network_size, mode):
-        prepared_data = []
-        image_id = 0
-        for tensor_image in tensor_images:
-            processed_images = padResize([tensor_image], network_size, mode)
-            for processed_image in processed_images:
-                prepared_data.extend(slice_and_separate_channels(processed_image, network_size, image_id))
-            image_id += 1
-        return prepared_data
-
-    def __getitem__(self, index):
-        item = self.prepared_data[index]
-        if self.transform:
-            transformed_image = self.transform(item['sub_image'])
-            
-            return {**item, 'sub_image': transformed_image}
-        return item
-
-    def __len__(self):
-        return len(self.prepared_data)
     
 
 def obtain_custom_input_dataset(tensor_images, config, mode='auto', transform = None):
